Rename `status_response` to `responses`
