+++ docs/examples/code/ping_as_java_and_bedrock_in_one_time.py
@@ -1,7 +1,7 @@
 import asyncio
 
 from mcstatus import BedrockServer, JavaServer
-from mcstatus.status_response import BedrockStatusResponse, JavaStatusResponse
+from mcstatus.responses import BedrockStatusResponse, JavaStatusResponse
 
 
 async def status(host: str) -> JavaStatusResponse | BedrockStatusResponse:
