+++ tests/test_bedrock_status.py
@@ -6,7 +6,7 @@ import pytest
 
 from mcstatus.address import Address
 from mcstatus.bedrock_status import BedrockServerStatus
-from mcstatus.status_response import BedrockStatusResponse
+from mcstatus.responses import BedrockStatusResponse
 
 
 def test_bedrock_response_is_expected_type():
